backend/requirements.txt
# Cherished Motion Lab - Backend Dependencies

# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
supabase==2.10.0
postgrest>=0.18.0,<0.19.0

# Authentication & Security
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# External APIs
openai==1.54.0
requests==2.32.3
httpx==0.27.2

# Video Processing
ffmpeg-python==0.2.0
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
